web/prowlbackend/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-cas-ng==3.5.9
django-cors-headers==2.3.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
idna==2.7
mysqlclient==1.3.13
python-cas==1.2.0
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
